+++ sdl2/ext/window.py
@@ -95,12 +95,12 @@ class Window(object):
         if flags is None:
             flags = self.DEFAULTFLAGS
 
+        self.window = None
         self._title = title
         self._position = position
         self._flags = flags
         self._size = size
 
-        self.window = None
         self.create()
 
     def __del__(self):
@@ -108,32 +108,40 @@ class Window(object):
         underlying SDL2 window."""
         self.close()
 
+    def _ensure_window(self, action):
+        if not self.window:
+            raise RuntimeError("Cannot {0} a closed window.".format(action))
+
     @property
     def position(self):
         """tuple: The (x, y) pixel coordinates of the top-left corner of the
         window.
 
         """
+        if not self.window:
+            return self._position
         x, y = c_int(), c_int()
         video.SDL_GetWindowPosition(self.window, byref(x), byref(y))
         return x.value, y.value
 
     @position.setter
     def position(self, value):
-        if not self.window:
-            raise SDLError("The window is not available")
-        video.SDL_SetWindowPosition(self.window, value[0], value[1])
+        if self.window:
+            video.SDL_SetWindowPosition(self.window, value[0], value[1])
         self._position = value[0], value[1]
 
     @property
     def title(self):
         """str: The title of the window."""
+        if not self.window:
+            return stringify(self._title, "utf-8")
         return stringify(video.SDL_GetWindowTitle(self.window), "utf-8")
 
     @title.setter
     def title(self, value):
-        title_bytes = utf8(value).encode('utf-8')
-        video.SDL_SetWindowTitle(self.window, title_bytes)
+        if self.window:
+            title_bytes = utf8(value).encode('utf-8')
+            video.SDL_SetWindowTitle(self.window, title_bytes)
         self._title = value
 
     @property
@@ -142,30 +150,33 @@ class Window(object):
         ``(width, height)``.
         
         """
+        if not self.window:
+            return self._size
         w, h = c_int(), c_int()
         video.SDL_GetWindowSize(self.window, byref(w), byref(h))
         return w.value, h.value
 
     @size.setter
     def size(self, value):
-        video.SDL_SetWindowSize(self.window, value[0], value[1])
         self._size = value[0], value[1]
+        if self.window:
+            video.SDL_SetWindowSize(self.window, value[0], value[1])
 
     def create(self):
         """Creates the window if it does not already exist."""
-        if self.window != None:
+        if self.window:
             return
         window = video.SDL_CreateWindow(utf8(self._title).encode('utf-8'),
                                         self._position[0], self._position[1],
                                         self._size[0], self._size[1],
                                         self._flags)
         if not window:
-            raise SDLError()
+            raise_sdl_err("creating the window")
         self.window = window.contents
 
     def open(self):
         """Creates and shows the window."""
-        if self.window is None:
+        if not self.window:
             self.create()
         self.show()
 
@@ -177,7 +188,7 @@ class Window(object):
         :meth:`create` method.
 
         """
-        if hasattr(self, "window") and self.window != None:
+        if self.window:
             video.SDL_DestroyWindow(self.window)
             self.window = None
 
@@ -187,6 +198,7 @@ class Window(object):
         If the window is already visible, this method does nothing.
 
         """
+        self._ensure_window("show")
         video.SDL_ShowWindow(self.window)
 
     def hide(self):
@@ -195,14 +207,17 @@ class Window(object):
         If the window is already hidden, this method does nothing.
         
         """
+        self._ensure_window("hide")
         video.SDL_HideWindow(self.window)
 
     def maximize(self):
         """Maximizes the window."""
+        self._ensure_window("maximize")
         video.SDL_MaximizeWindow(self.window)
 
     def minimize(self):
         """Minimizes the window into the system's dock or task bar."""
+        self._ensure_window("minimize")
         video.SDL_MinimizeWindow(self.window)
 
     def restore(self):
@@ -212,6 +227,7 @@ class Window(object):
         nothing.
 
         """
+        self._ensure_window("restore")
         video.SDL_RestoreWindow(self.window)
 
     def refresh(self):
@@ -222,6 +238,7 @@ class Window(object):
            modified using :meth:`get_surface`.
 
         """
+        self._ensure_window("refresh")
         video.SDL_UpdateWindowSurface(self.window)
 
     def get_surface(self):
@@ -245,7 +262,8 @@ class Window(object):
             :obj:`~sdl2.SDL_Surface`: The surface associated with the window.
 
         """
+        self._ensure_window("get the surface of")
         sf = video.SDL_GetWindowSurface(self.window)
         if not sf:
-            raise SDLError()
+            raise_sdl_err("getting the window surface")
         return sf.contents